👀

+++ Include/cpython/pyatomic_msc.h
@@ -1031,7 +1031,7 @@ static inline void
 _Py_atomic_store_ushort_release(unsigned short *obj, unsigned short value)
 {
 #if defined(_M_X64) || defined(_M_IX86)
-    *(short volatile *)obj = value;
+    *(unsigned short volatile *)obj = value;
 #elif defined(_M_ARM64)
     _Py_atomic_ASSERT_ARG_TYPE(unsigned __int16);
     __stlr16((unsigned __int16 volatile *)obj, (unsigned __int16)value);
@@ -1044,7 +1044,7 @@ static inline void
 _Py_atomic_store_uint_release(unsigned int *obj, unsigned int value)
 {
 #if defined(_M_X64) || defined(_M_IX86)
-    *(short volatile *)obj = value;
+    *(unsigned int volatile *)obj = value;
 #elif defined(_M_ARM64)
     _Py_atomic_ASSERT_ARG_TYPE(unsigned __int32);
     __stlr32((unsigned __int32 volatile *)obj, (unsigned __int32)value);
